feat(static-embeddings): add automodel
