+++ tests/test_fasttext_embeddings.py
@@ -5,11 +5,8 @@ import torch
 from flair.data import Sentence
 from torch.testing import assert_close as assert_close
 
-from embeddings.embedding.static.fasttext import (
-    KGR10FastTextConfig,
-    KGR10FastTextEmbedding,
-    SingleFileConfig,
-)
+from embeddings.embedding.static.config import SingleFileConfig
+from embeddings.embedding.static.fasttext import KGR10FastTextConfig, KGR10FastTextEmbedding
 from embeddings.utils.utils import import_from_string
